@@ -13,8 +13,8 @@
     "url": "https://github.com/martin8877/react-select-pre.git"
   },
   "dependencies": {
-    "classnames": "^2.2.4",
-    "prop-types": "^15.5.8",
+    "classnames": "^2.2.6",
+    "prop-types": "^15.7.2",
     "react-input-autosize": "^2.2.1"
   },
   "devDependencies": {
@@ -29,50 +29,52 @@
     "babel-preset-stage-0": "^6.24.1",
     "babel-register": "^6.26.0",
     "chai": "^4.1.2",
-    "coveralls": "^2.11.12",
+    "coveralls": "^3.0.9",
     "create-react-class": "^15.5.2",
-    "cross-env": "^5.0.5",
-    "css-loader": "^0.28.7",
-    "eslint": "^4.6.1",
-    "eslint-plugin-react": "^7.3.0",
-    "extract-text-webpack-plugin": "^3.0.0",
-    "gh-pages": "^1.1.0",
+    "cross-env": "^6.0.3",
+    "css-loader": "^3.2.0",
+    "eslint": "^6.7.1",
+    "eslint-plugin-react": "^7.16.0",
+    "extract-text-webpack-plugin": "^4.0.0-beta.0",
+    "gh-pages": "^2.1.1",
     "html-loader": "^0.5.1",
-    "html-webpack-plugin": "^2.30.1",
-    "husky": "^0.14.3",
+    "html-webpack-plugin": "^3.2.0",
+    "husky": "^3.1.0",
     "isomorphic-fetch": "^2.2.1",
     "istanbul": "^0.4.5",
-    "jsdom": "^9.12.0",
-    "less": "^2.7.2",
-    "less-loader": "^4.0.5",
+    "jsdom": "^15.2.1",
+    "less": "^3.10.3",
+    "less-loader": "^5.0.0",
     "less-plugin-clean-css": "^1.5.1",
-    "lint-staged": "^4.3.0",
-    "mocha": "^3.0.2",
+    "lint-staged": "^9.5.0",
+    "mocha": "^6.2.2",
     "nps": "^5.7.1",
     "nps-utils": "^1.3.0",
-    "nyc": "^11.1.0",
-    "react": "^15.5.0",
+    "nyc": "^14.1.1",
+    "react": "^16.3.1",
     "react-addons-shallow-compare": "^15.5.0",
-    "react-dom": "^15.5.0",
+    "react-dom": "^16.3.1",
     "react-gravatar": "^2.4.5",
-    "react-highlight-words": "^0.8.1",
-    "react-test-renderer": "^15.6.1",
+    "react-highlight-words": "^0.16.0",
+    "react-test-renderer": "^16.3.1",
     "react-virtualized": "^9.9.0",
     "react-virtualized-select": "^3.1.0",
-    "rollup": "^0.49.2",
-    "rollup-plugin-babel": "^3.0.2",
-    "rollup-plugin-commonjs": "^8.2.0",
-    "rollup-plugin-node-resolve": "^3.0.0",
-    "rollup-plugin-uglify": "^2.0.1",
-    "sinon": "^3.2.1",
-    "style-loader": "^0.18.2",
+    "rollup": "^1.27.5",
+    "rollup-plugin-babel": "^3.0.5",
+    "rollup-plugin-commonjs": "^10.1.0",
+    "rollup-plugin-node-resolve": "^5.2.0",
+    "rollup-plugin-uglify": "^6.0.3",
+    "sinon": "^7.5.0",
+    "style-loader": "^1.0.0",
     "uglify-es": "^3.0.28",
-    "unexpected": "^10.35.0",
-    "unexpected-dom": "^4.0.0",
-    "unexpected-react": "^4.1.0",
-    "unexpected-sinon": "^10.4.0",
-    "webpack": "^3.5.5",
-    "webpack-dev-server": "^2.7.1"
+    "unexpected": "^11.8.1",
+    "unexpected-dom": "^4.16.0",
+    "unexpected-htmllike": "^2.2.0",
+    "unexpected-react": "^5.0.2",
+    "unexpected-sinon": "^10.11.2",
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-server": "^3.9.0"
   },
   "peerDependencies": {
     "react": "^0.14.9 || ^15.3.0 || ^16.0.0-rc || ^16.0",
@@ -85,7 +87,7 @@
     "lint": "eslint .",
     "deploy": "cross-env NODE_ENV=production nps publish",
     "start": "webpack-dev-server --progress",
-    "test": "cross-env NODE_ENV=test mocha --compilers js:babel-core/register",
+    "test": "cross-env NODE_ENV=test mocha --require babel-core/register",
     "precommit": "lint-staged && yarn run test"
   },
   "files": [
