@@ -1,9 +1,8 @@
 module.exports = {
 	parserOptions: {
-		ecmaVersion: 6,
+		ecmaVersion: 2018,
 		sourceType: 'module',
 		ecmaFeatures: {
-			experimentalObjectRestSpread: true,
 			jsx: true,
 		},
 	},
